Czech translation fix
Hi,
I fixed some hard to understand terms, and hopefully improved the stylistic layout within display limits.

Note: I couldn't for the love of Pete figure out the encoding of the original file, neither Windows-1252 nor ANSI produced correct diacritics (out of many that was tested). So I re-encoded the file in ANSI, as specified in the file comments, in Notepad++, which should hopefully work.